+++ doc/release-notes.md
@@ -1,5 +1,6 @@
 # 1.3.2
 - cleanup of event folder
+- network policies added
 
 # 1.3.1
 - switchable kafka and rabbitmq broker added